🚚 Rename: user type 파일의 파일명을 user.types.ts로 변경
